% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATOMIC.R
\name{ATOMIC}
\alias{ATOMIC}
\title{Automated Imbalanced Classification}
\usage{
ATOMIC(
  form,
  train,
  nmodels = 20,
  metric = "F1",
  numCores = 1,
  modelPath = NULL,
  ...
)
}
\arguments{
\item{form}{A model formula}

\item{train}{Training data}

\item{nmodels}{Number of models to consider. Default is 20. Only such number of top-k models (based on internal validation performance with cross-validation methodology) will be tested.}

\item{metric}{Evaluation metric used for assessing the optimisation of predictive performance. Default is F1-Score}

\item{numCores}{number of cores for parallel computing}

\item{modelPath}{path to sysdata file containing details for meta-model}

\item{...}{Other parameters}
}
\value{
A predictive model containing the workflow (algorithm+resampling strategy) that are estimated to optimise the generalisation error.
}
\description{
A meta-learning-based AutoML approach to solving imbalanced classification tasks with resampling strategies.
}
\examples{
\dontrun{

library(mlbench)

data(PimaIndiansDiabetes)

form <- diabetes ~ .

atomic.m <- ATOMIC(form,PimaIndiansDiabetes)

}
}
