% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{attrWithOutliers}
\alias{attrWithOutliers}
\title{1. Attributes with outliers}
\usage{
attrWithOutliers(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A numeric value with the number of attributes where outliers are present
}
\description{
Attributes with outliers
}
