% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{classNames}
\alias{classNames}
\title{Returns the names of classes}
\usage{
classNames(form, ds)
}
\arguments{
\item{form}{A model formula}

\item{ds}{A data.frame object with the training data}
}
\value{
A vector
}
\description{
This function returns the names of classes in a nominal and binary target variable.
The vector is ordered: the majority class name is in the first position, and the minority class in the second.
}
