% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{ClassPD}
\alias{ClassPD}
\title{16. Percentual Difference by Class (PDC)
F1, F2 and F3}
\usage{
ClassPD(ds, form)
}
\arguments{
\item{ds}{A data set}

\item{form}{A model formula}
}
\value{
A vector of numerical values
}
\description{
Class-Based Percentual Differences
Applies all of the stats meta functions and calculates the percentual difference between values calculated based on the majority class and the minority class.
This is calculated as 100 * ( abs( f(X|Min) - f(X|Maj) ) / ( [ abs(f(X|Min)) + abs(f(X|Maj)) ] ) / 2 )
}
