% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{eval.func}
\alias{eval.func}
\title{Evaluation function to assess predictive ability of models}
\usage{
eval.func(form, train, test, preds)
}
\arguments{
\item{form}{A model formula}

\item{train}{A data.frame object with the training data}

\item{test}{A data.frame object with the test set}

\item{preds}{A vector with the predictions of a given model}
}
\value{
Evaluation Vector
}
\description{
Evaluation function to assess predictive ability of models
}
