% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{fscore}
\alias{fscore}
\title{Calculates the F1-Score for a pair set of true and predicted values}
\usage{
fscore(y_true, y_pred, positive = NULL)
}
\arguments{
\item{y_true}{Vector of size n with true values}

\item{y_pred}{Vector of size n with predicted values}

\item{positive}{String with the name of the positive class. If none is attributed it will select the least frequent class. Note: the distribution of classes in the population and the sample may be significantly different.}
}
\value{
F-Score value
}
\description{
Calculates the F1-Score for a pair set of true and predicted values
}
