% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafeatures.R
\name{getMetaFeatures}
\alias{getMetaFeatures}
\title{Meta-Feature Description}
\usage{
getMetaFeatures(ds, form, numCores = 1)
}
\arguments{
\item{ds}{A data set}

\item{form}{A model formula}

\item{numCores}{number of cores for parallel computing}
}
\value{
A vector of meta-features describing the data set
}
\description{
Provides anextensive description of a data set based on the extraction of meta-features
}
\examples{
\dontrun{

library(mlbench)

data(PimaIndiansDiabetes)

form <- diabetes ~ .

getMetaFeatures(PimaIndiansDiabetes,form)

}

}
