% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvalFramework.R
\name{kf_xval}
\alias{kf_xval}
\title{k-fold cross validation}
\usage{
kf_xval(
  x,
  form,
  nfolds,
  FUN,
  shuffle.rows = TRUE,
  average_results = FALSE,
  seedlock = FALSE,
  percTest = NULL,
  ...
)
}
\arguments{
\item{x}{data: embedded time series}

\item{form}{A model formula}

\item{nfolds}{no of folds}

\item{FUN}{function to apply to each iteration's train and validation. Typically
\strong{FUN} is a workflow where a predictive model is applied in a training set
and the model is evaluated in the validation set.}

\item{shuffle.rows}{Boolean for shuffling rows in cross-validation methodology. Default TRUE}

\item{average_results}{Boolean for return average results of metrics. Default FALSE}

\item{seedlock}{Random seed. Defaults to FALSE (appointing the seed 123)}

\item{percTest}{Percentage for holdout evaluation (generalisation). Default is NULL}

\item{...}{further parameters to \code{FUN}}
}
\description{
\code{kf_xval} applies a k-fold cross-validation methodology to a given data set.
}
