% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{landmarker.nb}
\alias{landmarker.nb}
\title{Naive Bayes Landmarkers}
\usage{
landmarker.nb(ds, form)
}
\arguments{
\item{ds}{A data set}

\item{form}{A model formula}
}
\value{
A vector of numerical values
}
\description{
Naive Bayes Landmarkers
}
