% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{landmarker.tree}
\alias{landmarker.tree}
\title{17. Landmarkers
Trees, NB}
\usage{
landmarker.tree(ds, form, maxdepth = 1)
}
\arguments{
\item{ds}{A data set}

\item{form}{A model formula}

\item{maxdepth}{The maximum depth for the tree}
}
\value{
A vector of numerical values
}
\description{
Decision-Tree Landmarkers
}
