% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{measOverlap}
\alias{measOverlap}
\title{15. Measures of Overlapping
F1, F2 and F3}
\usage{
measOverlap(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector of numeric values concerning Maximum Fisher's Discriminant Ratio (F1), the Volume of the Overlapping Region (F2), and the Maximum Individual Feature Efficiency (F3)
}
\description{
Measures of Overlapping
Includes Maximum Fishers' Discriminant Ratio (F1), the Volume of the Overlapping Region (F2), and the Maximum Individual Feature Efficiency (F3)
}
