% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{percDiff}
\alias{percDiff}
\title{Percentual Difference}
\usage{
percDiff(a, b)
}
\arguments{
\item{a}{Numerical value}

\item{b}{Numerical value}
}
\value{
Numerical value
}
\description{
This function calculates the percentual difference between two values (a - b)/b * 100
}
