% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATOMIC.R
\name{predict.atomic}
\alias{predict.atomic}
\title{Prediction method for the ATOMIC method}
\usage{
\method{predict}{atomic}(object, test, ...)
}
\arguments{
\item{object}{A prediction model}

\item{test}{Test data}

\item{...}{Additional parameters}
}
\value{
A vector of predicted values
}
\description{
Standard prediction method applied to an instance of the ATOMIC class
}
\examples{
library(mlbench)

data(PimaIndiansDiabetes)

ind <- sample(1:nrow(PimaIndiansDiabetes), 0.7*nrow(PimaIndiansDiabetes))

train <- PimaIndiansDiabetes[ind,]
test <- PimaIndiansDiabetes[-ind,]

form <- diabetes ~ .

atomic.m <- ATOMIC(form,PimaIndiansDiabetes)

preds <- predict(atomic.m, test)

}
