% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingStrategies.R
\name{rs.Adasyn}
\alias{rs.Adasyn}
\title{ADASYN}
\usage{
rs.Adasyn(form, train, beta = 0.5, ...)
}
\arguments{
\item{form}{A model formula}

\item{train}{A data.frame object with the training data}

\item{beta}{Balance level after synthetic case generation - 1 corresponds to fully balanced classes}

\item{...}{Additional parameters for the resampling strategy}
}
\value{
New data set
}
\description{
ADASYN
}
