% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingStrategies.R
\name{rs.ENN}
\alias{rs.ENN}
\title{Edited Nearest Neighbour}
\usage{
rs.ENN(form, train, k = 3, ...)
}
\arguments{
\item{form}{A model formula}

\item{train}{A data.frame object with the training data}

\item{k}{Number of nearest neighbours to consider (default is 3)}

\item{...}{Additional parameters for the resampling strategy}
}
\value{
New data set
}
\description{
Edited Nearest Neighbour
}
