% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingStrategies.R
\name{rs.RandOver}
\alias{rs.RandOver}
\title{Random Oversampling}
\usage{
rs.RandOver(form, train, ove.perc = 0.5, ...)
}
\arguments{
\item{form}{A model formula}

\item{train}{A data.frame object with the training data}

\item{ove.perc}{Oversampling percentage for the minority class (e.g. 20\% oversampling corresponds to 0.2)}

\item{...}{Additional parameters for the resampling strategy}
}
\value{
New data set
}
\description{
Random Oversampling
}
