% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingStrategies.R
\name{rs.RandUnder}
\alias{rs.RandUnder}
\title{Random Undersampling}
\usage{
rs.RandUnder(form, train, und.perc = 0.5, ...)
}
\arguments{
\item{form}{A model formula}

\item{train}{A data.frame object with the training data}

\item{und.perc}{Undersampling percentage for the majority class}

\item{...}{Additional parameters for the resampling strategy}
}
\value{
New data set
}
\description{
Random Undersampling
}
