% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingStrategies.R
\name{rs.SMOTE}
\alias{rs.SMOTE}
\title{Synthetic Minority Over-sampling Technique (SMOTE)}
\usage{
rs.SMOTE(form, train, und.perc = 1, ove.perc = 0, ...)
}
\arguments{
\item{form}{A model formula}

\item{train}{A data.frame object with the training data}

\item{und.perc}{Undersampling percentage for the majority class (default is 1)}

\item{ove.perc}{Oversampling percentage for the minority class (default is 1)}

\item{...}{Additional parameters for the resampling strategy}
}
\value{
New data set
}
\description{
Synthetic Minority Over-sampling Technique (SMOTE)
}
