% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingStrategies.R
\name{rs.TomekUnder}
\alias{rs.TomekUnder}
\title{Undersampling with Tomek links}
\usage{
rs.TomekUnder(form, train, rem = "maj", ...)
}
\arguments{
\item{form}{A model formula}

\item{train}{A data.frame object with the training data}

\item{rem}{Indicates if both parts of the Tomek link should be removed ("both") or solely the case from the majority class ("maj")}

\item{...}{Additional parameters for the resampling strategy}
}
\value{
New data set
}
\description{
Undersampling with Tomek links
}
