% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsCor_NumAttrs}
\alias{statsCor_NumAttrs}
\title{2. Correlation between Numeric Attributes
min, max, mean, sd, var, hist}
\usage{
statsCor_NumAttrs(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of values concerning correlation between numerical attributes,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on the Correlation between Numeric Attributes
}
