% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsCoV_NumAttrs}
\alias{statsCoV_NumAttrs}
\title{7. Coefficient of Variation
min, max, mean, sd, var, hist}
\usage{
statsCoV_NumAttrs(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of numerical attributes' coefficient of variation values,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on Coefficient of Variation of Numeric Attributes
}
