% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsEnt}
\alias{statsEnt}
\title{13. Entropy (Empirical) of Attributes
min, max, mean, sd, var, hist}
\usage{
statsEnt(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of attributes' entropy values,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on Entropy of Numeric Attributes
}
