% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsGKur_NumAttrs}
\alias{statsGKur_NumAttrs}
\title{4. Geary's Kurtosis of Numeric Attributes
min, max, mean, sd, var, hist}
\usage{
statsGKur_NumAttrs(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of numerical attributes' (Geary) kurtosis values,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on Geary's Kurtosis of Numeric Attributes
}
