% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsIQR_NumAttrs}
\alias{statsIQR_NumAttrs}
\title{3. Interquartile Range (IQR)
min, max, mean, sd, var, hist}
\usage{
statsIQR_NumAttrs(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of numerical attributes' IQR values,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on Interquartile Range of Numeric Attributes
}
