% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsMAS_NumAttrs}
\alias{statsMAS_NumAttrs}
\title{9. Maximum Asymmetry Score (MAS) between Numeric Attributes
min, max, mean, sd, var, hist}
\usage{
statsMAS_NumAttrs(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of numerical attributes' MAS score,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on Maximum Asymmetry Score (MAS) between Numeric Attributes
}
