% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsMCN_NumAttrs}
\alias{statsMCN_NumAttrs}
\title{11. Minimum Cell Number (MCN) between Numeric Attributes
min, max, mean, sd, var, hist}
\usage{
statsMCN_NumAttrs(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of numerical attributes' MCN score,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on Minimum Cell Number (MCN) between Numeric Attributes
}
