% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsMEV_NumAttrs}
\alias{statsMEV_NumAttrs}
\title{10. Maximum Edge Value (MEV) between Numeric Attributes
min, max, mean, sd, var, hist}
\usage{
statsMEV_NumAttrs(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of numerical attributes' MEV score,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on Maximum Edge Value (MEV) between Numeric Attributes
}
