% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsMIC_NumAttrs}
\alias{statsMIC_NumAttrs}
\title{8. Maximal Information Coefficient (MIC) between Numeric Attributes
min, max, mean, sd, var, hist}
\usage{
statsMIC_NumAttrs(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of numerical attributes' MIC score,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on Maximal Information Coefficient (MIC) between Numeric Attributes
}
