% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsMuI}
\alias{statsMuI}
\title{14. Mutual Information of Attributes
min, max, mean, sd, var, hist}
\usage{
statsMuI(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of attributes' mutual information values,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on Mutual Information of Attributes
}
