% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsSkew_NumAttrs}
\alias{statsSkew_NumAttrs}
\title{6. Skewness of Numeric Attributes
min, max, mean, sd, var, hist}
\usage{
statsSkew_NumAttrs(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of numerical attributes' (Pearson) skewness values,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on Skewness of Numeric Attributes
}
