% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metafunctions.R
\name{statsTIC_NumAttrs}
\alias{statsTIC_NumAttrs}
\title{12. Total Information Coefficient (TIC) between Numeric Attributes
min, max, mean, sd, var, hist}
\usage{
statsTIC_NumAttrs(ds, tgt)
}
\arguments{
\item{ds}{A data set}

\item{tgt}{The index of the target variable}
}
\value{
A vector with the minimum, maximum, mean, standard deviation and variance of numerical attributes' TIC score,
and a vector with the cardinality of values after binning (10 bins)
}
\description{
Statistics on Total Information Coefficient (TIC) between Numeric Attributes
}
