% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflows.R
\name{wf.NaiveBayes}
\alias{wf.NaiveBayes}
\title{Naive Bayes (Landmarker)}
\usage{
wf.NaiveBayes(form, train, test, ...)
}
\arguments{
\item{form}{A model formula}

\item{train}{Train data}

\item{test}{Test data}

\item{...}{Additional parameters for rpart.control}
}
\value{
Evaluation Vector
}
\description{
Naive Bayes (Landmarker)
}
