% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflows.R
\name{wf.RandomForest}
\alias{wf.RandomForest}
\title{Workflow with Random Forest learning algorithm}
\usage{
wf.RandomForest(form, train, test, ntrees = 500, rstrategy = NULL,
  rs.pars = list(), return.model = FALSE, ...)
}
\arguments{
\item{form}{A model formula}

\item{train}{Train data}

\item{test}{Test data}

\item{ntrees}{Number of trees (algorithm parameter)}

\item{rstrategy}{Name of Resampling Strategy}

\item{rs.pars}{Parameters of Resampling Strategy}

\item{return.model}{If TRUE, this function will return the model and not its predictive performance. Default FALSE.}

\item{...}{Other parameters for the learning algorithm}
}
\value{
Evaluation Vector
}
\description{
Workflow with Random Forest learning algorithm
}
