% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workflows.R
\name{wf.Tree}
\alias{wf.Tree}
\title{Decision Tree (Landmarker)}
\usage{
wf.Tree(form, train, test, maxdepth = 4, ...)
}
\arguments{
\item{form}{A model formula}

\item{train}{Train data}

\item{test}{Test data}

\item{maxdepth}{Depth parameter for tree-based learning}

\item{...}{Additional parameters for rpart.control}
}
\value{
Evaluation vector
}
\description{
Decision Tree (Landmarker)
}
